feat(core): new redesigned cloud page
