+++ src/components/common/Quotes.vue
@@ -36,8 +36,10 @@ const activeQuote = ref(0)
         <div class="content">
             <img v-if="quotes[activeQuote].bigLogoSrc" :src="quotes[activeQuote].bigLogoSrc" class="big-logo" />
             <div class="text" :class="{'text-center': !quotes[activeQuote].bigLogoSrc}">
-                <blockquote>{{ quotes[activeQuote].quote }}</blockquote>
-                <span v-html="quotes[activeQuote].author" />
+                <h1 class="quote-title" data-usal="fade-r">
+                    <blockquote class="quote-content">{{ quotes[activeQuote].quote }}</blockquote>
+                </h1>
+                <p v-html="quotes[activeQuote].author" data-usal="fade-l"/>
                 <div class="small-logo" v-if="quotes.length === 1" v-html="quotes[activeQuote].smallLogoSvg" />
             </div>
         </div>
@@ -59,15 +61,18 @@ const activeQuote = ref(0)
 
 <style lang="scss" scoped>
 @import "~/assets/styles/variable";
+
 .wrapper {
     background-color: var(--ks-background-purple-light);
     color: var(--ks-content-primary);
+
     .tabs {
         display: flex;
         justify-content: center;
         border-bottom: 1px solid var(--ks-border-secondary);
         margin: 0 auto;
-        .tab{
+
+        .tab {
             background: none;
             border: none;
             cursor: pointer;
@@ -78,12 +83,15 @@ const activeQuote = ref(0)
             justify-content: center;
             color: var(--ks-content-primary);
             flex: 1;
+
             :deep(svg) {
                 height: 1.5rem;
                 fill: var(--ks-content-primary);
                 transition: fill 0.3s ease;
             }
-            &[aria-selected='true'], &:hover {
+
+            &[aria-selected='true'],
+            &:hover {
                 background-color: var(--ks-background-purple-hover);
                 border-bottom: 1px solid var(--ks-border-active);
                 margin-bottom: -1px;
@@ -91,53 +99,79 @@ const activeQuote = ref(0)
         }
     }
 
-    .content{
+    .content {
         padding: 4rem;
         display: flex;
         flex-direction: column;
         align-items: center;
         gap: 2rem;
+        max-width: 1200px;
+        margin-inline: auto;
         .big-logo {
             // guarantee square proportions
             aspect-ratio: 1 / 1;
             object-fit: contain;
             max-width: 80%;
         }
+
         @include media-breakpoint-up(lg) {
             flex-direction: row;
+
             .big-logo {
                 max-width: 35%;
             }
         }
-        .text-center {
-            text-align: center;
-        }
-        blockquote{
-            font-size: 28px;
-            &::after, &::before {
-                color: var(--ks-content-color-highlight);
-                font-size: 2rem;
-                line-height: 2.5rem;
-                margin: .2rem;
-                vertical-align: bottom;
-                margin-left: 0.5rem;
+
+        .text {
+            display: flex;
+            flex-direction: column;
+            gap: 2rem;
+            align-items: center;
+
+            .quote-title {
+                margin: 0;
+                color: var(--ks-content-primary);
+                font-weight: normal;
+                line-height: 1.3;
+
+                .quote-content {
+                    display: inline;
+                    padding: 0;
+                    margin: 0;
+
+                    &::before,
+                    &::after {
+                        color: var(--ks-content-color-highlight);
+                    }
+
+                    &::before {
+                        content: "“";
+                        margin-right: 0.25rem;
+                    }
+
+                    &::after {
+                        content: "”";
+                        margin-left: 0.25rem;
+                    }
+                }
             }
-            &::before {
-                content: '“';
+
+            p {
+                margin: 0;
+
+                @include media-breakpoint-up(lg) {
+                    font-size: 1.125rem;
+                }
+
+                font-family: $font-family-monospace;
+                color: var(--ks-content-secondary);
             }
-            &::after {
-                content: '”';
-            }
-        }
-        > div > span{
-            font-family: var(--font-family-jetbrains-mono);
-            font-size: 1rem;
-        }
-        .small-logo{
-            margin-top: 2rem;
-            :deep(svg) {
-                height: 2rem;
-                fill: var(--ks-content-primary);
+
+            .small-logo {
+                :deep(svg) {
+                    height: 2rem;
+                    fill: var(--ks-content-primary);
+                }
             }
         }
     }
@@ -150,14 +184,17 @@ const activeQuote = ref(0)
         width: 70%;
         margin: 0 auto;
         padding: 1rem 0;
+
         .kpi {
             display: flex;
             flex-direction: column;
             align-items: center;
             flex: 1;
+
             strong {
                 font-size: 1.5rem;
             }
+
             span {
                 font-size: 0.7rem;
                 color: var(--ks-content-secondary);
@@ -172,23 +209,29 @@ const activeQuote = ref(0)
         align-items: center;
         font-size: .8rem;
         padding: 1rem;
-        a, button {
+
+        a,
+        button {
             color: var(--ks-content-primary);
             border: 1px solid var(--ks-border-primary-inverse);
             border-radius: 2px;
             text-decoration: none;
             font-size: .5rem;
             padding: .15rem .6rem;
+
             &:hover {
                 background: var(--ks-background-purple-hover);
             }
         }
-        a{
+
+        a {
             font-weight: bold;
         }
+
         button {
             cursor: pointer;
             background: none;
+
             &:disabled {
                 color: var(--ks-content-secondary);
                 cursor: not-allowed;
