+++ src/pages/cloud.astro
@@ -1,118 +1,23 @@
 ---
 import Layout from "~/components/layout.astro"
+import Quotes from "~/components/common/Quotes.vue"
 import Header from "~/components/cloud/Header.astro"
-import Features from "~/components/cloud/Features.vue"
+import foundationLogo from "~/components/cloud/assets/foundation-direct-logo.svg?raw"
 
 const title = "Kestra Cloud Early Adopter Program"
 const description =
     "Be among the first to experience Kestra Cloud with exclusive access before general availability"
+
+const quotes = [
+    {
+        quote: "Kestra Cloud has been a pivotal part of giving us flexibility and scalability we need to pull off complex processes we do at Foundation Direct.",
+        author: "Michael Heidner - SVP of Analytics and Business Intelligence",
+        smallLogoSvg: foundationLogo,
+    },
+]
 ---
 
 <Layout {title} {description}>
-    <div class="cloud">
-        <Header />
-        <Features client:idle />
-    </div>
+    <Header />
+    <Quotes {quotes} client:load />
 </Layout>
-
-<style scoped lang="scss">
-    @import "~/assets/styles/variable";
-
-    .cloud {
-        overflow: hidden;
-        &::after {
-            content: "";
-            position: absolute;
-            top: 0;
-            right: 0;
-            width: 100%;
-            height: 100%;
-            background-image: url("~/components/cloud/images/gradient_top.png");
-            background-size: contain;
-            background-repeat: no-repeat;
-            background-position: top right;
-            transition: all 0.3s ease;
-            z-index: -1;
-        }
-
-        @media (max-width: 575.98px) {
-            &::after {
-                right: auto;
-                left: 0;
-                transform: scaleX(-1);
-            }
-        }
-
-        @media (max-width: 767.98px) {
-            &::after {
-                right: auto;
-                left: 0;
-                transform: scaleX(-1);
-            }
-        }
-
-        @media (max-width: 991.98px) {
-            &::after {
-                right: auto;
-                left: 0;
-                transform: scaleX(-1);
-            }
-        }
-
-        & :global(h1),
-        & :global(h2),
-        & :global(h3),
-        & :global(h4),
-        & :global(h5),
-        & :global(h6) {
-            color: var(--ks-content-primary);
-            font-weight: 600;
-        }
-
-        & :global(h1) {
-            font-size: $font-size-3xl;
-        }
-
-        & :global(h2) {
-            font-size: $font-size-2xl;
-        }
-
-        & :global(h3) {
-            font-size: $font-size-2xl;
-        }
-
-        & :global(h4) {
-            font-size: $font-size-xl;
-        }
-
-        & :global(h5) {
-            font-size: $font-size-lg;
-        }
-
-        & :global(h6) {
-            font-size: $font-size-md;
-            font-weight: 600;
-        }
-
-        & :global(p) {
-            font-size: 14px;
-            color: var(--ks-content-primary);
-        }
-
-        & :global(.form) {
-            & h4,
-            & p,
-            & div {
-                color: var(--ks-content-primary);
-            }
-
-            & div {
-                &.success {
-                    padding-top: 100px;
-                }
-
-                font-size: $font-size-sm;
-            }
-        }
-    }
-</style>
